added cleanup code
