+++ tests/protractor/conf.js
@@ -70,7 +70,7 @@ const setupUser = () => {
 
 exports.config = {
   seleniumAddress: 'http://localhost:4444/wd/hub',
-  specs: ['e2e/**/submit-delivery*.js'],
+  specs: ['e2e/**/submit-delivery-*.js'],
   framework: 'jasmine2',
   capabilities: {
     // browserName: 'chrome'
